+++ petitRADTRANS/__file_conversion.py
@@ -2951,7 +2951,7 @@ def format2petitradtrans(load_function, opacities_directory: str, natural_abunda
                          save_line_by_line: bool = True, line_by_line_wavelength_boundaries: np.ndarray = None,
                          standard_line_by_line_wavelength_boundaries: np.ndarray = None, rewrite: bool = False,
                          use_legacy_correlated_k_wavenumbers_sampling: bool = False):
-    """Convert opacities loaded with the specied load function into petitRADTRANS line-by-line and correlated-k
+    """Convert opacities loaded with the specified load function into petitRADTRANS line-by-line and correlated-k
     opacities.
 
     Args:
@@ -2965,17 +2965,21 @@ def format2petitradtrans(load_function, opacities_directory: str, natural_abunda
                 - file_extension: see opacity_files_extension below
                 - molmass: molar mass of the species (automatically set)
                 - wavelength_file: see spectral_dimension_file below
-                - wavenumbers_petitradtrans: the petitRADTRANS wavenumber grid (automatically set)
+                - wavenumbers_petitradtrans_line_by_line: the petitRADTRANS wavenumber grid for line-by-line opacities
+                  (automatically set)
                 - save_line_by_line: see save_line_by_line below
                 - rebin: see rebin below
-                - selection: indices corresponding to the wavenumbers to be exctracted
+                - selection: indices corresponding to the wavenumbers to be extracted
             Not all of the above arguments have to be used.
             The function must output the following, in that order:
-                - opacities: (cm2/molecule) the opacities
-                - opacities_line_by_line: (cm2/molecule), the opacities, interpolated to wavenumbers_petitradtrans
-                - wavenumbers: (cm-1) the wavenumbers corresponding to opacities
-                - pressure: (bar) the pressure of the opacities
-                - temperature: (K) the temperature of the opacities
+                - cross_sections: (cm2/molecule) the cross-sections
+                - cross_sections_line_by_line: (cm2/molecule), the cross-sections,
+                    interpolated to wavenumbers_petitradtrans_line_by_line
+                - wavenumbers: (cm-1) the wavenumbers corresponding to cross-sections
+                - pressure: (bar) the pressure of the cross-sections
+                - temperature: (K) the temperature of the cross-sections
+            The cross_sections, cross_sections_line_by_line and wavenumbers must be returned in increasing wavenumber
+            order.
         opacities_directory:
             Directory in which the opacity files are stored.
         natural_abundance:
@@ -3126,11 +3130,16 @@ def format2petitradtrans(load_function, opacities_directory: str, natural_abunda
             file_extension=opacity_files_extension,
             wavelength_file=spectral_dimension_file,
             molmass=molmass,
-            wavenumbers_petitradtrans=wavenumbers_petitradtrans,
+            wavenumbers_petitradtrans_line_by_line=wavenumbers_line_by_line,
             save_line_by_line=save_line_by_line,
             rebin=rebin,
             selection=selection
         )
+
+        # Raise error if the wavenumbers are not sorted in increasing order
+        if not np.all(np.diff(wavenumbers) > 0):
+            raise ValueError("wavenumbers (and cross-sections) returned by external opacity loading function must be "
+                             "sorted in increasing order")
 
         if i == 0:
             __wavenumbers = wavenumbers
@@ -3596,13 +3605,13 @@ def line_by_line_opacities_dat2h5(directory, output_name, molmass, doi,
         __remove_files([directory])
 
 
-def load_dace(file, file_extension, molmass, wavelength_file=None, wavenumbers_petitradtrans=None,
+def load_dace(file, file_extension, molmass, wavelength_file=None, wavenumbers_petitradtrans_line_by_line=None,
               save_line_by_line=False, rebin=True, selection=None):
     """Read a DACE opacity file."""
     import struct
 
-    if wavenumbers_petitradtrans is None:
-        raise TypeError("missing 1 required argument: 'wavenumbers_petitradtrans'")
+    if wavenumbers_petitradtrans_line_by_line is None:
+        raise TypeError("missing 1 required argument: 'wavenumbers_petitradtrans_line_by_line'")
 
     if wavelength_file is not None:
         warnings.warn("Dace opacities does not require a wavelength file\n"
@@ -3647,29 +3656,33 @@ def load_dace(file, file_extension, molmass, wavelength_file=None, wavenumbers_p
     # Handle insufficient wavelength coverage
     d_wavenumbers = np.mean(np.diff(wavenumbers))
 
-    if wavenumber_start > wavenumbers_petitradtrans[selection[0]]:
-        warnings.warn(f"wavenumber coverage of converted opacities does not extend to "
-                      f"the requested wavelength lower value "
-                      f"({wavenumber_start}  cm-1 > {wavenumbers_petitradtrans[selection[0]]} cm-1)"
-                      f"\nOpacities set to 0 within {wavenumbers_petitradtrans[selection[0]]}--{wavenumber_start} cm-1")
+    if wavenumber_start > wavenumbers_petitradtrans_line_by_line[selection[0]]:
+        warnings.warn(
+          f"wavenumber coverage of converted opacities does not extend to "
+          f"the requested wavelength lower value "
+          f"({wavenumber_start}  cm-1 > {wavenumbers_petitradtrans_line_by_line[selection[0]]} cm-1)"
+          f"\nOpacities set to 0 within {wavenumbers_petitradtrans_line_by_line[selection[0]]}--{wavenumber_start} cm-1"
+        )
         wavenumbers = np.insert(
-            wavenumbers, 0, [wavenumbers_petitradtrans[selection[0]], wavenumber_start - d_wavenumbers]
+            wavenumbers, 0, [wavenumbers_petitradtrans_line_by_line[selection[0]], wavenumber_start - d_wavenumbers]
         )
         opacities = np.insert(opacities, 0, [0, 0])
 
-    if wavenumber_end < wavenumbers_petitradtrans[selection[1]]:
-        warnings.warn(f"wavenumber coverage of converted opacities does not extend to "
-                      f"the requested wavelength higher value "
-                      f"({wavenumber_end} cm-1 < {wavenumbers_petitradtrans[selection[1]]} cm-1)"
-                      f"\nOpacities set to 0 within {wavenumber_end}--{wavenumbers_petitradtrans[selection[1]]} cm-1")
+    if wavenumber_end < wavenumbers_petitradtrans_line_by_line[selection[1]]:
+        warnings.warn(
+            f"wavenumber coverage of converted opacities does not extend to "
+            f"the requested wavelength higher value "
+            f"({wavenumber_end} cm-1 < {wavenumbers_petitradtrans_line_by_line[selection[1]]} cm-1)"
+            f"\nOpacities set to 0 within {wavenumber_end}--{wavenumbers_petitradtrans_line_by_line[selection[1]]} cm-1"
+        )
         wavenumbers = np.concatenate(
-            [wavenumbers, [wavenumber_end + d_wavenumbers, wavenumbers_petitradtrans[selection[1]]]]
+            [wavenumbers, [wavenumber_end + d_wavenumbers, wavenumbers_petitradtrans_line_by_line[selection[1]]]]
         )
         opacities = np.concatenate([opacities, [0, 0]])
 
     # Interpolate the Dace calculation to that grid
     sig_interp = interp1d(wavenumbers, opacities)
-    sigmas_prt = sig_interp(wavenumbers_petitradtrans[selection[0]:selection[1]])
+    sigmas_prt = sig_interp(wavenumbers_petitradtrans_line_by_line[selection[0]:selection[1]])
 
     # Check if interp values are below 0 or NaN
     if np.any(np.less(sigmas_prt, 0)):
@@ -3691,13 +3704,13 @@ def load_dace(file, file_extension, molmass, wavelength_file=None, wavenumbers_p
     return opacities, opacities_line_by_line, wavenumbers, pressure, temperature
 
 
-def load_exocross(file, file_extension, molmass=None, wavelength_file=None, wavenumbers_petitradtrans=None,
+def load_exocross(file, file_extension, molmass=None, wavelength_file=None, wavenumbers_petitradtrans_line_by_line=None,
                   save_line_by_line=False, rebin=True, selection=None):
     if wavelength_file is None:
         raise TypeError("missing required argument 'wavelength_file'")
 
-    if wavenumbers_petitradtrans is None:
-        raise TypeError("missing required argument 'wavenumbers_petitradtrans'")
+    if wavenumbers_petitradtrans_line_by_line is None:
+        raise TypeError("missing required argument 'wavenumbers_petitradtrans_line_by_line'")
 
     if molmass is not None:
         pass  # silent warning
@@ -3738,7 +3751,7 @@ def load_exocross(file, file_extension, molmass=None, wavelength_file=None, wave
         # Interpolate the ExoCross calculation to that grid
         print(" Interpolating...")
         sig_interp = interp1d(wavenumbers, opacities)
-        sigmas_prt = sig_interp(wavenumbers_petitradtrans)
+        sigmas_prt = sig_interp(wavenumbers_petitradtrans_line_by_line)
     else:
         n_lines = finput.count_file_line_number_sequential(file)
         sigma = finput.load_all_line_by_line_opacities_sequential(file, n_lines)