+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to petitRADTRANS will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com)
 and this project adheres to [Semantic Versioning](http://semver.org).
 
-## [3.1.0a28] - 2024-07-30
+## [3.1.0a32] - 2024-08-06
 ### Added
 - Possibility to generate mock input data for input == output retrievals, using the exact same format as the input data.
 - Possibility to run pRT's retrieval model with emcee (base implementation with less functionality than the full retrieval package, i.e., no plotting support for result analysis)
@@ -30,11 +30,15 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 ### Changed
 - Future: parameter `emission_geometry` is canonically renamed `irradiation_geometry`. Parameter `emission_geometry` will be deprecated in version 4.0.0.
 - Clarified a bit the documentation on the `SpectralModel` retrieval framework.
+- Requested input and output parameter names for externally provided function to load opacities for `format2petitradtrans`: since cm^2 should be returned it should be called cross-sections, not opacities.
 
 ### Removed
 - Unused test functions.
 
 ### Fixed
+- Bug in function `retrieval.plot_spectra()` when plotting the best-fit spectrum together with `radtrans_grid=True`.
+- Bug in function `calculate_transit_radii()` when `return_opacities=True`.
+- Function `format2petitradtrans` applied the incorrect pRT wavelength grid to the lbl opacity conversion.
 - Function `rebin_spectrum_bin` incorrectly handling overlapping bins.
 - Crash when unpickling `LockedDict` objects.
 - Crash when loading unspecified source opacities with different spectral info than the default opacity file and multiple files with that spectral info exist.